added example chart

+++ lib/repository/crypto_coin/models/crypto_coin_details.dart
@@ -23,7 +23,7 @@ class CryptoCoinDetails extends Equatable {
   final double sentimentVotesDownPercentage;
   final int watchlistPortfolioUsers;
   final int? marketCapRank;
-  final int? coingeckoRank;
+  final double? coingeckoRank;
   final double coingeckoScore;
   final double developerScore;
   final double communityScore;
